backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy>=1.24.0,<2.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless>=4.8.0
scipy>=1.10.0
scikit-image>=0.21.0
Pillow==10.1.0
bcrypt==4.1.1
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
apscheduler==3.10.4
easyocr==1.7.0
pytesseract==0.3.10

